pyproject.toml
[project]
name = "clairvoyance"
version = "0.1.0"
description = "Personal finance data engineering pipeline"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.25.0",
    "pdfplumber>=0.11.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "google-cloud-storage>=2.16.0",
    "google-cloud-bigquery>=3.20.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
include = ["ingestion*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
